rename grain generation
