+++ protobuf/Proto.Cluster.CodeGen.Tests/bar.proto
