+++ protobuf/Proto.Cluster.CodeGen.Tests/foo.proto
