+++ protobuf/Proto.Cluster.CodeGen/Template.cs
@@ -3,7 +3,7 @@
 //      Copyright (C) 2015-2020 Asynkron AB All rights reserved
 // </copyright>
 // -----------------------------------------------------------------------
-namespace ProtoBuf
+namespace Proto.Cluster.CodeGen
 {
     public static class Template
     {
