+++ protobuf/Proto.Cluster.CodeGen/buildMultiTargeting/ProtoGrainGenerator.props
@@ -3,5 +3,5 @@
   when it targets multiple frameworks, e.g. <TargetFrameworks>net46;netcoreapp1.0</TargetFrameworks>
 -->
 <Project>
-  <Import Project="..\build\ProtoGrainGenerator.props" />
+  <Import Project="..\build\Proto.Cluster.CodeGen.props" />
 </Project>
