+++ protobuf/Proto.Cluster.CodeGen/model/ProtoMessage.cs
@@ -3,9 +3,7 @@
 //      Copyright (C) 2015-2021 Asynkron AB All rights reserved
 // </copyright>
 // -----------------------------------------------------------------------
-using System.Collections.Generic;
-
-namespace ProtoBuf
+namespace Proto.Cluster.CodeGen.model
 {
     public class ProtoMessage
     {
