+++ protobuf/Proto.Cluster.CodeGen/model/ProtoService.cs
@@ -3,7 +3,7 @@
 //      Copyright (C) 2015-2021 Asynkron AB All rights reserved
 // </copyright>
 // -----------------------------------------------------------------------
-namespace ProtoBuf
+namespace Proto.Cluster.CodeGen.model
 {
     public class ProtoService
     {
